+++ camelid/tests/test_camelidenv.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 '''Unit tests for camelid run environment.'''
 
-from camelid.run_camelid import CamelidEnv
+from camelid.run import CamelidEnv
 
 
 def test_camelidenv():
